+++ src/main/kotlin/com/viartemev/requestmapper/annotations/UrlFormatter.kt
@@ -1,0 +1,6 @@
+package com.viartemev.requestmapper.annotations
+
+interface UrlFormatter {
+
+    fun format(classMapping: String, methodMapping: String, param: String = ""): String
+}
